requirements.txt
google-adk[vertexai]
google-cloud-aiplatform>=1.36.0
